Improve Virtual Tour Length

+++ TwoOpt.hpp
@@ -20,6 +20,7 @@
 #include <map>
 #include <queue>
 #include <algorithm>
+#include <limits>
 
 class City{
 private:
@@ -82,7 +83,7 @@ private:
     std::vector<City*> testTour;
     void exchangeCities(int, int);
     double calculateTourLength(std::vector<City*>);
-    double calcVirtTourLength(int, int);
+    int calcVirtTourLength(int, int);
     
 public:
     Trip(std::vector<City>*);
